requirements.txt
jupyter
pandas
numpy
langchain
langchain-openai
huggingface_hub
openai
pinecone-client 
chromadb
streamlit
python-dotenv